backend/requirements.txt
aioredis==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
dask==2024.12.0
entrypoints==0.4
fabio==2024.9.0
fastapi==0.115.6
fastapi-cache==0.1.0
fsspec==2024.10.0
h11==0.14.0
h5py==3.12.1
hdf5plugin==5.0.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
json-merge-patch==0.2
jsonpatch==1.33
jsonpointer==3.0.0
locket==1.0.0
lxml==5.3.0
msgpack==1.1.0
numpy==2.2.0
orjson==3.10.12
packaging==24.2
partd==1.4.2
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
tiled==0.1.0b12
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
